- clarifacation de l'affichage des frames - methode de modif des probas, passage de int a double pour les probas - ajout de la continuité des mouvements sur Forward et Backward

+++ IAstageStateV2/src/AttackState.java
@@ -16,7 +16,7 @@ public abstract class AttackState extends State {
 		int probaTot = 100;
 		boolean exit = false;
 		
-		for(int i : probabilities)
+		for(double i : probabilities)
 		{
 			if (choice >= probaTot - i)
 			{
